+++ TindahanNatin.Server/Models/IAuditable.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace TindahanNatin.Server.Models;
+
+public interface IAuditable
+{
+    DateTimeOffset CreatedAt { get; set; }
+    DateTimeOffset UpdatedAt { get; set; }
+}