+++ TindahanNatin.Server/Models/Shelf.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace TindahanNatin.Server.Models;
 
-public class Shelf
+public class Shelf : ISoftDelete, IAuditable
 {
     public Guid Id { get; set; } = Guid.NewGuid();
     public string Name { get; set; } = string.Empty;