+++ tindahan_natin/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0+14] - 2026-05-31
+
+### Added
+- Add Share to quick actions in dashboard
+- Move Share and Log out to main settings screen
+
+### Fixed
+- Performance improvements
+
 ## [1.0.0+13] - 2026-05-30
 
 ### Added